spatial grid optimizations

+++ server/java/shared/src/main/java/io/gamemachine/client/api/ClientGrid.java
@@ -174,7 +174,7 @@ public class ClientGrid {
 		}
 	}
 
-	public Boolean set(String id, float x, float y, float z, EntityType entityType) {
+	public Boolean set(String id, int x, int y, int z, EntityType entityType) {
 		TrackData trackData = new TrackData();
 		trackData.setId(id);
 		trackData.setX(x);
